add jump and crouch

+++ includes/game/entities/monster_entity.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   monster_entity.h                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lbenard <lbenard@student.42.fr>            +#+  +:+       +#+        */
+/*   By: mribouch <mribouch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/19 19:07:35 by lbenard           #+#    #+#             */
-/*   Updated: 2020/08/10 21:20:56 by lbenard          ###   ########.fr       */
+/*   Updated: 2020/08/12 01:44:34 by mribouch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -74,7 +74,7 @@ t_result		init_monster_entity(t_monster_entity *const self,
 void			monster_entity_update(t_monster_entity *const self);
 int				distance_monster_player(t_monster_entity *self);
 void			update_agro_distance(t_monster_entity *self);
-int				get_distance_vec(t_vec3f source, t_vec3f target);
+float				get_distance_vec(t_vec3f source, t_vec3f target);
 int				get_orientate_sprite(t_monster_entity *self);
 void			a_star_attack(t_monster_entity *self,
 					int distance,
